# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Forge CI/CD API
  description: Self-hosted CI/CD control plane. Array responses, error envelope and in-process rate limits follow docs/contracts/API_CONTRACT.md and docs/API.md current compatibility mode.
  license:
    name: FerrPOINT Proprietary Source-Available Evaluation License v1.0
    url: https://github.com/FerrPOINT/CI-CD/blob/main/LICENSE
  version: 0.1.0
paths:
...
  /api/v1/projects:
    get:
      tags:
      - projects
      operationId: list_projects
      parameters:
      - name: limit
        in: query
        description: Max items (1..=200, default 50).
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      - name: offset
        in: query
        description: Offset (default 0).
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      tags:
      - projects
      operationId: create_project
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProject'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: validation error
  /api/v1/projects/{project_id}:
    get:
      tags:
      - projects
      operationId: get_project
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: ''
    delete:
      tags:
      - projects
      operationId: delete_project
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: ''
        '404':
          description: ''
    patch:
      tags:
      - projects
      operationId: update_project
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProject'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: ''
...
  /api/v1/projects/{project_id}/memberships:
    get:
      tags:
      - memberships
      operationId: list_project_memberships
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectMembership'
        '404':
          description: ''
    post:
      tags:
      - memberships
      operationId: upsert_project_membership
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectMembershipInput'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMembership'
        '400':
          description: ''
        '404':
          description: ''
  /api/v1/projects/{project_id}/memberships/{user_id}:
    delete:
      tags:
      - memberships
      operationId: delete_project_membership
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: ''
        '404':
          description: ''
        '409':
          description: ''
...
components:
  schemas:
...
    CreateProject:
      type: object
      required:
      - name
      - repository_url
      properties:
        default_branch:
          type:
          - string
          - 'null'
        name:
          type: string
        repository_url:
          type: string
        tenant_id:
          type:
          - string
          - 'null'
          format: uuid
          description: 'Target-model tenancy (AUTHORIZATION): optional during backfill.'
...
    Project:
      type: object
      required:
      - id
      - name
      - repository_url
      - default_branch
      - created_at
      properties:
        created_at:
          type: string
          format: date-time
        default_branch:
          type: string
        id:
          type: string
          format: uuid
        max_running_jobs:
          type:
          - integer
          - 'null'
          format: int32
          description: 'K4.3 dispatch cap: null = unlimited.'
        name:
          type: string
        repository_url:
          type: string
    ProjectMembership:
      type: object
      required:
      - project_id
      - user_id
      - username
      - user_enabled
      - role
      - created_at
      - updated_at
      properties:
        created_at:
          type: string
          format: date-time
        project_id:
          type: string
          format: uuid
        role:
          type: string
        updated_at:
          type: string
          format: date-time
        user_enabled:
          type: boolean
        user_id:
          type: string
          format: uuid
        username:
          type: string
    ProjectMembershipInput:
      type: object
      required:
      - user_id
      - role
      properties:
        role:
          type: string
        user_id:
          type: string
          format: uuid
...
    UpdateProject:
      type: object
      properties:
        default_branch:
          type:
          - string
          - 'null'
        max_running_jobs:
          type:
          - integer
          - 'null'
          format: int32
          description: |-
            K4.3 dispatch fairness: max concurrently active-leased jobs for this
            project (>=1). Omit to keep; null clears the limit.
        name:
          type:
          - string
          - 'null'
        repository_url:
          type:
          - string
          - 'null'
...